pyproject.toml
[tool.poetry]
name = "cr-search"
version = "0.1.0"
description = ""
authors = ["Lukas Winkler <git@lw1.at>"]

[tool.poetry.dependencies]
python = "^3.8"
peewee = "^3.13.1"
PyMySQL = "^0.9.3"
srt = "^3.0.0"
sonic-client = {path = "python-sonic-client"}
psycopg2 = "^2.8.4"
youtube-dl = "^2020.2.16"
spacy = "^2.2.3"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz"}
progress = "^1.5"
flask = "^1.1.1"
sentry_sdk = "^0.14.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
